pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ediscovery-agents"
version = "0.1.0"
description = "eDiscovery Time Entry Analysis System — local AI agents powered by Google ADK and Ollama"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.25.0",
    "litellm",
    "pandas",
    "openpyxl",
    "pyarrow",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
ediscovery = "ediscovery_agents.__main__:main"

[tool.setuptools.packages.find]
include = ["ediscovery_agents*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
